refactor(api-client): what belongs to the client is named by the client
Unwrapping an eden reply and naming the client's own type are both facts about the client rather than about whoever calls it, so neither is the cli's to declare.
