+++ packages/cli/src/lib/api.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from 'bun:test';
-import { authHeaders, describeFailure } from '#lib/api.ts';
+import { authHeaders } from '#lib/api.ts';
 
 const API_URL = 'http://localhost:3000';
 
@@ -25,30 +25,4 @@ test('a token issued by another api is not sent to this one', () => {
 // nothing to authenticate with.
 test('being signed out builds a client rather than refusing to', () => {
   expect(authHeaders({ baseUrl: API_URL, credentials: null })).toEqual({});
-});
-
-test('a refusal is reported as the api answering, not as this program deciding', () => {
-  const failure = Object.assign(new Error('[object Object]'), {
-    status: 404,
-    value: { error: 'Not Found' },
-  });
-
-  expect(describeFailure(failure)).toBe('The api answered 404: Not Found');
-});
-
-test('a body that names no error is still worth repeating', () => {
-  const failure = Object.assign(new Error('nope'), { status: 500, value: 'nope' });
-
-  expect(describeFailure(failure)).toBe('The api answered 500: nope');
-});
-
-// Eden reports a request it never sent as a 503 of its own, and saying the api answered it would
-// blame a server that was never reached.
-test('an api that could not be reached is not one that answered', () => {
-  const failure = Object.assign(new Error('unreachable'), {
-    status: 503,
-    value: new TypeError('Unable to connect'),
-  });
-
-  expect(describeFailure(failure)).toBe('Unable to connect');
 });
