+++ packages/cli/src/lib/errors.ts
@@ -1,10 +1,3 @@
-export class ApiError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = 'ApiError';
-  }
-}
-
 /** Something wrong with what was typed, decided here rather than by sending it and being told. */
 export class UsageError extends Error {
   constructor(message: string) {
